requirements.txt
colorama
colorspacious
ipykernel
ipython
loguru
mock
numpy
numpydoc
pandas
patsy
plotly
python-dotenv
pytz
PyVirtualDisplay
PyYAML
regex
requests
scikit-learn
seaborn
# selenium
sklearn
statsmodels
termcolor
terminado